% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table_interface.R
\name{data_table_interface}
\alias{data_table_interface}
\title{Create a \code{data.table} interface around a proxy object}
\usage{
data_table_interface(table_proxy)
}
\arguments{
\item{table_proxy}{Object generated with method \code{table_proxy()} that provides an interface
to an encapsulated remote table.}
}
\value{
A \code{datatableinterface} object that exposes a \code{data.table} interface to the encapsulated
\code{table_proxy} object.
}
\description{
This method defines a \code{data.table} interface around a specific object. That object should
define all the generics needed for supporting the interface.
}
