% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table_proxy_class.R
\name{data_table_proxy}
\alias{data_table_proxy}
\title{Create a \code{data.table} interface around a proxy object}
\usage{
data_table_proxy(remote_proxy)
}
\arguments{
\item{remote_proxy}{A object with a custom class. That class should implement the generics
required for a data.table proxy}
}
\value{
A \code{remote_proxy} object with a \code{data.table} interface
}
\description{
This method defines a \code{data.table} interface around a specific object. That object should
define all the generics needed for supporting the interface.
}
