% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_table_fst.R
\name{remote_table_fst}
\alias{remote_table_fst}
\title{Proxy object to a fst file}
\usage{
remote_table_fst(path, old_format = FALSE)
}
\arguments{
\item{path}{Path to a \emph{fst} file}

\item{old_format}{Use \code{old_format = TRUE} when referencing a \emph{fst} file that was created
with a `fst`` package version lower than 0.8.0}
}
\value{
An object of class \code{remote_table_fst}
}
\description{
This method creates a proxy object to a \emph{fst} file. With this object and the accompanying
methods, it's possible to retrieve data from a \emph{fst} file using complex filters, selections
and orderings. The \code{fst_proxy} class serves as the interface between the \code{fst} package and
the \code{data.table} external interface. In the future, it can be re-used as the proxy between
\code{fst} and other interfaces (e.g. \code{fstplyr}, \code{SQL})
}
