% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table_interface.R
\name{[.datatableinterface}
\alias{[.datatableinterface}
\title{a data.table interface}
\usage{
\method{[}{datatableinterface}(x, i, j, by, keyby, with, nomatch, mult,
  roll, rollends, which, .SDcols, verbose = FALSE, collect = FALSE,
  allow.cartesian, drop, on)
}
\arguments{
\item{x}{a data.table interface object}

\item{i}{data.table's 'i' argument}

\item{j}{data.table's 'j' argument}

\item{by}{not implemented at this point}

\item{keyby}{not implemented at this point}

\item{with}{not implemented at this point}

\item{nomatch}{not implemented at this point}

\item{mult}{not implemented at this point}

\item{roll}{not implemented at this point}

\item{rollends}{not implemented at this point}

\item{which}{not implemented at this point}

\item{.SDcols}{not implemented at this point}

\item{verbose}{If TRUE, status and information messages are returned to the console.}

\item{collect}{If TRUE, the data.table interface operation will return a data.table
to memory}

\item{allow.cartesian}{not implemented at this point}

\item{drop}{not implemented at this point}

\item{on}{not implemented at this point}
}
\description{
a data.table interface
}
