% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_proxy.R
\name{table_proxy}
\alias{table_proxy}
\title{A proxy for a remote table}
\usage{
table_proxy(remote_table)
}
\arguments{
\item{remote_table}{A object with a custom class. That class should implement the generics
required for a data.table proxy}
}
\value{
A \code{table_proxy} object that can be used a a virtual table.
}
\description{
This method defines a \code{table_proxy} object that encapsulates a \code{remote_table}.
A \code{table_proxy} is a virtual construct that emulates a full table. The actual
column data is kept in the physical table as long as possible to keep interaction
with the \code{table_proxy} as fast as possible. Operations on a \code{table_proxy} object
yield a new \code{table_proxy} object with (possibly) modified meta-data to emulate the
result.
}
