% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_proxy_methods.R
\name{table_proxy_read_full}
\alias{table_proxy_read_full}
\title{Read the all data selected by the table_proxy into memory, and put into the specified order}
\usage{
table_proxy_read_full(tbl_proxy, col_names = NULL)
}
\arguments{
\item{tbl_proxy}{a \code{table_proxy} object}

\item{col_names}{names of the columns to retrieve. If NULL, all columns in the tbl_proxy will be returned}
}
\value{
a data.frame object with the requested data
}
\description{
For now, all rows in the full range are read in, and the uneeded rows are discarded.
}
