% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_proxy_methods.R
\name{table_proxy_read_range}
\alias{table_proxy_read_range}
\title{Read a subset of the proxy table to memory}
\usage{
table_proxy_read_range(tbl_proxy, from_row, to_row, col_names = NULL)
}
\arguments{
\item{tbl_proxy}{a \code{table_proxy} object}

\item{from_row}{start row (of table_proxy)}

\item{to_row}{end row (of table_proxy)}

\item{colnames}{names of the columns that you want to retrieve}
}
\value{
a data.frame object with the requested data
}
\description{
Read a subset of the proxy table to memory
}
