% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_proxy_methods.R
\name{table_proxy_select_row_mask}
\alias{table_proxy_select_row_mask}
\title{Apply a binary row-selection operation on the current table_proxy state.
This operation will not change the slice map ordering}
\usage{
table_proxy_select_row_mask(tbl_proxy, i)
}
\arguments{
\item{tbl_proxy}{a table proxy object}

\item{i}{a logical vector with the required rows set to TRUE}
}
\value{
a table proxy object with the new state
}
\description{
Apply a binary row-selection operation on the current table_proxy state.
This operation will not change the slice map ordering
}
