% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_proxy_methods.R
\name{table_proxy_select_rows}
\alias{table_proxy_select_rows}
\title{Apply a row-selection operation on the current table_proxy state
This operation could change the slice map ordering}
\usage{
table_proxy_select_rows(tbl_proxy, i)
}
\arguments{
\item{tbl_proxy}{a table proxy object}

\item{i}{an integer vector with the selected rows}
}
\value{
a table proxy object with the new state
}
\description{
Apply a row-selection operation on the current table_proxy state
This operation could change the slice map ordering
}
