% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_proxy_methods.R
\name{table_proxy_transform}
\alias{table_proxy_transform}
\title{Apply a column transformation operation on the current table_proxy state}
\usage{
table_proxy_transform(tbl_proxy, colexps)
}
\arguments{
\item{tbl_proxy}{a table proxy object}

\item{j}{column expressions}
}
\value{
a table proxy object with the new state
}
\description{
Currently this is intended to be used with the j argument to a datatableinterface
object, but it could just as well be used to implement mutate or transmute for a
dplyr interface.
}
