% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_table.R
\name{fst_table}
\alias{fst_table}
\title{Access a fst file like a data table}
\usage{
fst_table(path, old_format = FALSE)
}
\value{
An object of class \code{fsttable}
}
\description{
These functions permit using a dataset stored in a fst file much like a regular
data table.
}
